Simp2

+++ sum.py
@@ -1,6 +1,6 @@
 #input from user
-num1 = float(input("enter first number: "))
-num2 = float(input("enter second number: "))
+num1 = int(input("enter first number: "))
+num2 = int(input("enter second number: "))
 #calculate the sum of two numbers
 sum = num1 + num2
 #display result
